+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## Unreleased
+## Version 0.0.3 (development)
 
+- Exit with error code if log file does not exist
 - Set `--azureml` automatically when running on Azure ML
 
 ## Version 0.0.2 (development)